Add login_required decorator for homepage
Added an index page function which routes the user to the homepage in case he is already logged in
The `home` function routes the user to the relevant homepage rendering function. The `home` function has a `login_required` decorator because only logged in users should be able to access the homepage.
Logged out users who will try to access the homepage will get redircted to the landing page.