Add function to get the matching apartments for seeker
`get_matched_apartments` returns an ordered QuerySet of apartments which match the preferences of the user, ordered by the matching score creating by the ranking algorithm.

Note: `get_all_relevant_apartments` is a temporary function which was written in order to create this feature. Once the search feature is created, this function must be deleted, as we'll use the function in the search feature.